Ecommerce_Projects/requirements.txt
appdirs==1.4.4
asgiref==3.5.2
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cargo==0.3
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.6
comtypes==1.2.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cssselect2==0.7.0
cycler==0.11.0
debuggingbook==1.1.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dill==0.3.6
distlib==0.3.6
dj-database-url==0.5.0
Django==4.0
django-cas-ng==4.3.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-extensions==3.2.3
django-filter==23.5
django-groups-manager==1.3.0
django-guardian==2.4.0
django-hosts==5.1
django-js-asset==2.1.0
django-mama-cas==2.5.0
django-mptt==0.15.0
django-phonenumber-field==6.3.0
django-rest-auth==0.9.5
django-simple-sso==1.1.0
django-soet==1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.2.0
docutils==0.20.1
drf-yasg==1.21.4
easyplotly==0.1.3
enforce==0.3.4
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
fonttools==4.39.3
fuzzingbook==1.1
geo21==1.0.0
gitdb==4.0.10
GitPython==3.1.30
graphviz==0.20.1
h11==0.14.0
idna==3.4
importlib-metadata==7.0.1
inflection==0.5.1
ipython==8.12.0
itsdangerous==0.24
itypes==1.2.0
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonfield==3.1.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-pygments==0.2.2
keyring==24.3.0
kiwisolver==1.4.4
lizard==1.17.10
lxml==4.9.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mathutils @ file:///C:/Users/4/Desktop/Python--Library/mathutils/mathutils
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
more-itertools==10.2.0
multiprocess==0.70.14
mypy==1.2.0
mypy-extensions==1.0.0
mysql==0.0.3
mysql-client==0.0.1
mysql-connector==2.2.9
mysqlclient==2.2.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
networkx==3.1
nh3==0.2.15
notedown==1.5.1
npm==0.1.1
numpy==1.24.2
oauthlib==3.2.1
optional-django==0.1.0
outcome==1.2.0
packaging==21.3
pandas==1.5.3
pandoc-attributes==0.1.7
pandocfilters==1.5.0
parso==0.8.3
pep517==0.13.1
pickleshare==0.7.5
pika==1.3.0
Pillow==9.3.0
pkginfo==1.9.6
platformdirs==3.4.0
plotly==5.14.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.9
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pydot==1.4.2
PyDriller==2.4.1
Pygments==2.15.1
PyJWT==1.7.1
PyMySQL==1.1.0
pyotp==2.7.0
pyparsing==2.4.7
pypiwin32==223
pyproject-toml==0.0.10
pyrsistent==0.19.3
PySocks==1.7.1
python-cas==1.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-magic==0.4.27
python3-openid==3.2.0
pyttsx3==2.90
pytz==2022.6
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.0.2
readme-renderer==42.0
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.9.0
shapely==2.0.1
showast==0.2.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2
stripe==5.0.0
svglib==1.5.1
tenacity==8.2.2
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.3.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
twine==4.0.2
types-Markdown==3.4.2.7
types-pytz==2023.3.0.0
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.0.31
virtualenvwrapper-win==1.2.7
vonage==3.1.0
wcwidth==0.2.6
webencodings==0.5.1
webservices==0.7
whitenoise==5.2.0
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.2.0
yapf==0.33.0
z3-solver==4.12.1.0
zipp==3.17.0
